# perf: Avoid redundant copying of arrays in scan->filter->join