--- a/native/core/Cargo.toml
+++ b/native/core/Cargo.toml
@@ -68,8 +68,10 @@ paste = "1.0.14"
 datafusion-common = { workspace = true }
 datafusion = {  workspace = true }
 datafusion-expr = {  workspace = true }
+datafusion-execution = {  workspace = true }
 datafusion-physical-expr-common = {  workspace = true }
 datafusion-physical-expr = {  workspace = true }
+datafusion-physical-plan = {  workspace = true }
 once_cell = "1.18.0"
 regex = { workspace = true }
 crc32fast = "1.3.2"