--- a/native/core/src/execution/datafusion/operators/mod.rs
+++ b/native/core/src/execution/datafusion/operators/mod.rs
@@ -16,3 +16,4 @@
 // under the License.
 
 pub mod expand;
+pub(crate) mod filter;