## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes https://github.com/apache/datafusion-comet/issues/757

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

Improve performance of joins by removing some redundant copying of arrays for join inputs.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

This PR adds a new `CometFilterExec` which is a copy of DataFusion's `FilterExec` with one small change to ensure that input arrays are never emitted without copying:

```rust
// BEGIN HACK: this is the only modification to the DataFusion version
// we ensure that we never return the original arrays
let filtered_batch = if filtered_batch.num_rows() == batch.num_rows() {
    RecordBatch::try_new(
        filtered_batch.schema(),
        filtered_batch
            .columns()
            .iter()
            .map(|array| copy_array(array))
            .collect(),
    )?
} else {
    filtered_batch
};
// END HACK
```

In `planner.rs` we have this check for join inputs:

```rust
let left = if can_reuse_input_batch(&left) {
    Arc::new(CopyExec::new(left))
} else {
    left
};
```

Because `can_reuse_input_batch` returns `false` for `CometFilterExec`, we are longer creating the `CopyExec`.

## How are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

Existing tests