@andygrove 2024-08-02
I'm not seeing any performance improvement so far

## Before
```
OpenJDK 64-Bit Server VM 11.0.24+8-post-Ubuntu-1ubuntu322.04 on Linux 6.5.0-41-generic
AMD Ryzen 9 7950X3D 16-Core Processor
TPCDS Micro Benchmarks:                   Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
join_filtered_scans                                 596            696         115         24.4          41.1       1.0X
join_filtered_scans: Comet (Scan, Exec)            1162           1224          93         12.5          80.1       0.5X
```

## After

```
OpenJDK 64-Bit Server VM 11.0.24+8-post-Ubuntu-1ubuntu322.04 on Linux 6.5.0-41-generic
AMD Ryzen 9 7950X3D 16-Core Processor
TPCDS Micro Benchmarks:                   Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------
join_filtered_scans                                 574            688         129         25.3          39.6       1.0X
join_filtered_scans: Comet (Scan, Exec)            1181           1231          65         12.3          81.5       0.5X
```